# BiLock bench and speed up