Three patches:

* bench
* `BiLockAcquired.unlock` improvement
* `BiLockAcquire.poll` improvement

First patch adds two `BiLock` benches: contended and uncontended `lock`/`unlock`.

Second and third patch optimize `BiLockAcquired::unlock`. Uncontended `lock`/`unlock` bench speeds up by about 30%.

After first patch (bench added):

```
test contended   ... bench:     246,615 ns/iter (+/- 46,844)
test lock_unlock ... bench:     153,971 ns/iter (+/- 39,179)
```

After third patch:

```
test contended   ... bench:     187,539 ns/iter (+/- 41,419)
test lock_unlock ... bench:      94,046 ns/iter (+/- 18,137)
```