@alexcrichton 2017-06-01
Thanks! Could the unsafe logic in https://github.com/alexcrichton/futures-rs/pull/492/commits/264083aed0720b0f1684830be7d978fde2aa8aa4 be replaced with an `Option`? It seems like it may have a similar speedup perhaps?