@stepancheg 2017-06-01
Yes, it is the same with `Option`. Updated third patch.